+++ Rdmp.Core.Tests/CommandExecution/ExecuteCommandConfirmLogsTests.cs
@@ -30,7 +30,7 @@ namespace Rdmp.Core.Tests.CommandExecution
             lm.CreateNewLoggingTaskIfNotExists("GGG");
 
             var cmd = new ExecuteCommandConfirmLogs(new ThrowImmediatelyActivator(RepositoryLocator), lmd);
-            var ex = Assert.Throws<ExecuteCommandConfirmLogs.LogsNotConfirmedException>(()=>cmd.Execute());
+            var ex = Assert.Throws<LogsNotConfirmedException>(()=>cmd.Execute());
 
             Assert.AreEqual("There are no log entries for MyLmd", ex.Message);
 
@@ -70,12 +70,12 @@ namespace Rdmp.Core.Tests.CommandExecution
 
             var lm = new LogManager(lmd.GetDistinctLoggingDatabase());
             lm.CreateNewLoggingTaskIfNotExists("FFF");
-            var logEntry = lm.CreateDataLoadInfo("FFF", "pack o' cards", "going down gambling", null, true);
-            
-            // we have created log entry but it did not have an end time.  This is a sad entry because it never completed
 
+            // we have created log entry but it did not have an end time.  This is a sad entry because it never completed
+            lm.CreateDataLoadInfo("FFF", "pack o' cards", "going down gambling", null, true);
+            
             var cmd = new ExecuteCommandConfirmLogs(new ThrowImmediatelyActivator(RepositoryLocator), lmd);
-            var ex = Assert.Throws<ExecuteCommandConfirmLogs.LogsNotConfirmedException>(() => cmd.Execute());
+            var ex = Assert.Throws<LogsNotConfirmedException>(() => cmd.Execute());
 
             StringAssert.IsMatch("Latest logs for MyLmd .* indicate that it did not complete", ex.Message);
         }
@@ -94,7 +94,7 @@ namespace Rdmp.Core.Tests.CommandExecution
             logEntry.LogFatalError("vegas", "we lost it all on a pair of deuces");
 
             var cmd = new ExecuteCommandConfirmLogs(new ThrowImmediatelyActivator(RepositoryLocator), lmd);
-            var ex = Assert.Throws<ExecuteCommandConfirmLogs.LogsNotConfirmedException>(() => cmd.Execute());
+            var ex = Assert.Throws<LogsNotConfirmedException>(() => cmd.Execute());
 
             StringAssert.IsMatch("Latest logs for MyLmd .* indicate that it failed", ex.Message);
 
@@ -122,7 +122,7 @@ namespace Rdmp.Core.Tests.CommandExecution
 
             // but we want this to have finished in the last second
             var cmd = new ExecuteCommandConfirmLogs(new ThrowImmediatelyActivator(RepositoryLocator), lmd,"00:00:01");
-            var ex = Assert.Throws<ExecuteCommandConfirmLogs.LogsNotConfirmedException>(() => cmd.Execute());
+            var ex = Assert.Throws<LogsNotConfirmedException>(() => cmd.Execute());
 
             StringAssert.IsMatch("Latest logged activity for MyLmd is .*.  This is older than the requested date threshold:.*", ex.Message);
         }