+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Console gui context menu now shows compatible commands from plugins
+
 ## [7.0.1] - 2021-10-27
 
 ### Changed