+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- When syncing table columns with the database, the full column (including table name) is displayed in the proposed fix (previously only the column name was displayed).
+
 ## [7.0.1] - 2021-10-27
 
 ### Changed