+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added the `ConfirmLogs` command for verifying if a task is failing (e.g. a DLE run)
+
 ## [7.0.1] - 2021-10-27
 
 ### Changed