+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added the `ConfirmLogs` command for verifying if a task is failing (e.g. a DLE run)
+- Added the 'ConfirmLogs' command for verifying if a task is failing (e.g. a DLE run)
 
 ### Changed