+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed 'package downgrade' dependencies issue with `HIC.RDMP.Plugin.UI`
 - Fixed log viewer total time display in logs view when task ran for > 24 hours.
+- Fixed not implemented Exception when using username/password authentication and viewing [CohortIdentificationConfiguration] SQL
 
 ### Added