+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed 'package downgrade' dependencies issue with `HIC.RDMP.Plugin.UI`
 - Fixed log viewer total time display in logs view when task ran for > 24 hours.
 - Fixed not implemented Exception when using username/password authentication and viewing [CohortIdentificationConfiguration] SQL
+- Fixed missing 'add sql file process task' in DLE load stage right click context menus
+
 
 ### Added